更新功能 添加依赖

+++ PixivAss/Client_Util.cs
@@ -4,11 +4,11 @@ using System.IO;
 using System.Threading.Tasks;
 using System.Net;
 using System.Net.Http;
+using System.Text;
 using HtmlAgilityPack;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using PixivAss.Data;
-using IDManLib;
 
 namespace PixivAss
 {
@@ -51,28 +51,30 @@ namespace PixivAss
             return String.Format("{0}_p{1}.{2}", illust.id, page, ext);
         }
 
-        public string GetDownloadDir(Illust illust)
-        {
-            if (illust.bookmarked && illust.bookmarkPrivate)
-                return this.formal_private_dir;
-            else if (illust.bookmarked && !illust.bookmarkPrivate)
-                return this.formal_public_dir;
-            return this.tmp_dir;
-        }
+       
         //将指定图片下载到本地
         //如已存在则先删除
-        public bool DownloadIllustForce(ICIDMLinkTransmitter idm,string id, string url, string dir, string file_name)
+        public async Task<bool> DownloadIllustForceAria2(string url, string dir, string file_name)
         {
-            string path = dir + "/" + file_name;
             try
             {
+                if (string.IsNullOrEmpty(url))
+                    throw new ArgumentNullException("url");
+                string path = dir + "/" + file_name;
                 if (File.Exists(path))
                     File.Delete(path);
-                Console.WriteLine("Begin:" + file_name);
-                string referer = String.Format("{0}member_illust.php?mode=medium&illust_id={1}", base_url, id);
-                //referer = String.Format("https://www.pixiv.net/artworks/{0}",id);
-                //0x01:不确认，0x02:稍后下载
-                idm.SendLinkToIDM(url, referer, cookie_server.cookie, "", "", "", dir, file_name, 0x01 | 0x02);
+                /*id必须有，值可以随便填
+                 * 虽然url是数组但是并不能一次下载多个
+                 * token(rpc secret)和其它参数的格式不一样
+                 * 失败时RequesttAria2Async会直接抛出异常所以此处无需验证返回的json
+                */
+                //dir = "E:/test/2";
+                var data = String.Format("{{\"jsonrpc\": \"2.0\",\"id\":\"PixivAss\",\"method\": \"aria2.addUri\"," +
+                                    "\"params\": [\"token:{0}\",[\"{1}\"],{{\"dir\":\"{2}\",\"out\":\"{3}\""+
+                                    //",\"Cookie\":\"{4}\""+
+                                    "}}]}}",
+                                    aria2_rpc_secret,url,dir,file_name,cookie_server.cookie);
+                await RequesttAria2Async(data);                
             }
             catch (Exception e)
             {
@@ -80,6 +82,30 @@ namespace PixivAss
                 return false;
             }
             return true;
+        }
+        public async Task<bool> QueryAria2Status()
+        {
+            try
+            {
+                var data = String.Format("{{\"jsonrpc\": \"2.0\",\"id\":\"PixivAss\",\"method\": \"aria2.getGlobalStat\"," +
+                                    "\"params\": [\"token:{0}\"]}}",
+                                    aria2_rpc_secret);
+                var ret=(JObject)JsonConvert.DeserializeObject(await RequesttAria2Async(data));
+                var result=ret.Value<JObject>("result");
+                float speed = (result.Value<Int32>("downloadSpeed") >> 10)/1024.0f;
+                int active = result.Value<Int32>("numActive");
+                int waiting = result.Value<Int32>("numWaiting");
+                int done = result.Value<Int32>("numStoppedTotal");
+
+                Console.WriteLine("Aria2 Download Status:{0}MB/s of {1}(Running)/{2}(Waiting)/{3}(Done) Task",
+                    speed,active,waiting,done);
+                return active == waiting && active == 0;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                throw;
+            }
         }
         /*
          * 网络请求
@@ -89,7 +115,22 @@ namespace PixivAss
             if (!response.IsSuccessStatusCode)
                 throw new Exception("HTTP Not Success");
         }
-        public async Task<string> RequestAsync(string url, Uri referer)
+        public async Task<string> RequesttAria2Async(String data)
+        {
+            try
+            {
+                HttpResponseMessage response = await httpClient.PostAsync(aria2_rpc_addr, new StringContent(data));
+                CheckStatusCode(response);
+                return await response.Content.ReadAsStringAsync();
+            }
+            catch (Exception e)
+            {
+                string msg = e.Message;//e.InnerException.InnerException.Message;
+                Console.WriteLine("Request Aria RPC Fail :" + msg);
+                throw;
+            }
+        }
+        public async Task<string> RequestPixivAsyncGet(string url, Uri referer)
         {
             int try_ct = 5;
             while (true)
@@ -103,7 +144,7 @@ namespace PixivAss
                         throw new ArgumentException("Not SSL");
                     httpClient.DefaultRequestHeaders.Referrer = referer;
                     HttpResponseMessage response = await httpClient.GetAsync(url);
-                    //作品已删除
+                    //可能是作品已删除，此时仍然返回结果
                     if (response.StatusCode == HttpStatusCode.NotFound)
                         return await response.Content.ReadAsStringAsync();
                     //未知错误
@@ -139,33 +180,69 @@ namespace PixivAss
                 }
             }
         }
+        public async Task<string> RequestPixivAsyncPost(string url,Uri referer,string data)
+        {
+            for (int try_ct = 5; try_ct >= 0; --try_ct)
+                try
+                {
+                    if (string.IsNullOrEmpty(url))
+                        throw new ArgumentNullException("url");
+                    if (!url.StartsWith("https"))
+                        throw new ArgumentException("Not SSL");
+                    httpClient.DefaultRequestHeaders.Referrer = referer;
+                    HttpResponseMessage response = await httpClient.PostAsync(url, new StringContent(data,Encoding.UTF8,"application/json"));
+                    var ret = await response.Content.ReadAsStringAsync();
+                    Console.WriteLine(response.StatusCode.ToString()+":"+ret);
+                    CheckStatusCode(response);
+                    return ret;
+                }
+                catch (Exception e)
+                {
+                    Console.WriteLine(e.Message + "Re Try " + try_ct.ToString() + " On :" + url);
+                    if (try_ct == 0)
+                        throw;
+                }
+            return "";
+        }
         public async Task<JObject> RequestJsonAsync(string url, string referer)
         {
-            return (JObject)JsonConvert.DeserializeObject(await RequestAsync(url, new Uri(referer)));
+            var r = await RequestPixivAsyncGet(url, new Uri(referer));
+            return (JObject)JsonConvert.DeserializeObject(r);
         }
         public async Task<HtmlDocument> RequestHtmlAsync(string url, string referer)
         {
             var doc = new HtmlDocument();
-            var ret = await RequestAsync(url, new Uri(referer));
+            var ret = await RequestPixivAsyncGet(url, new Uri(referer));
             doc.LoadHtml(ret);
             return doc;
         }
-        public async Task<User> RequestUserAsync(string userId)
+        public async Task<User> RequestUserAsync(int userId)
         {
-            string url = String.Format("{0}ajax/user/{1}/profile/top", base_url, userId);
-            string referer = String.Format("{0}member_illust.php?id={1}", base_url, user_id);
-            JObject ret = await RequestJsonAsync(url, referer);
-            if (ret.Value<Boolean>("error"))
-                throw new Exception("Get User Fail");
-            var userName = ret.GetValue("body").Value<JObject>("extraData").Value<JObject>("meta").Value<string>("title");
-            if (ret.GetValue("body").Value<JObject>("illusts").Count > 0)
-                foreach (var illustId in ret.GetValue("body").Value<JObject>("illusts"))
-                {
-                    var illust = await RequestIllustAsync(illustId.Key);
-                    userName = illust.userName;
-                    break;
-                }
-            return new User(userId, userName, false);
+            try
+            {
+                string url = String.Format("{0}ajax/user/{1}/profile/top", base_url, userId);
+                string referer = String.Format("{0}member_illust.php?id={1}", base_url, user_id);
+                JObject ret = await RequestJsonAsync(url, referer);
+                if (ret.Value<Boolean>("error"))
+                    throw new Exception("Get User Fail");
+                var body = ret.Value<JObject>("body");
+                var userName = body.Value<JObject>("extraData").Value<JObject>("meta").Value<string>("title");
+                foreach (var type in new List<string>{ "illusts","manga"})
+                    if(body.GetValue(type)!=null)
+                        if(body.GetValue(type).Type==JTokenType.Object)//必须判断类型，因为空的时候是个空array而非object，很迷
+                            foreach (var illustId in body.Value<JObject>(type))
+                            {
+                                var illust = await RequestIllustAsync(Int32.Parse(illustId.Key));
+                                userName = illust.userName;
+                                break;
+                            }
+                return new User(userId, userName, false,false);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                throw;
+            }            
         }
         public async Task<int> RequestFollowedUserCount()
         {
@@ -176,8 +253,54 @@ namespace PixivAss
                 throw new Exception("Get Bookmark Fail");
             return ret.GetValue("body").Value<int>("total");
         }
-       
-        public async Task<List<string>> RequestSearchPage(string word, int page, bool text_mode)
+
+        public async Task<bool> PushBookmark(bool bookmarked,int illust_id,bool pub, Int64 bookmark_id=-1)
+        {
+            if(bookmarked)
+            {
+                var ret = await RequestPixivAsyncPost(String.Format("{0}ajax/illusts/bookmarks/add", base_url),
+                                            new Uri(String.Format("{0}artworks/{1}", base_url, illust_id)),
+                                            String.Format("{{\"illust_id\": \"{0}\", \"restrict\": {1}, \"comment\": \"\", \"tags\": []}}", illust_id, pub ? 0 : 1));
+                var json = (JObject)JsonConvert.DeserializeObject(ret);
+                if (json != null)
+                    if (!json.Value<Boolean>("error"))
+                        return true;
+                return false;
+            }
+            else
+            {
+                if (bookmark_id < 0)
+                    throw new ArgumentNullException("There must be a Bookmark ID when delete a bookmark");
+                var ret=await RequestPixivAsyncPost(String.Format("{0}bookmark_setting.php", base_url),
+                    new Uri(String.Format("{0}bookmark_add.php?type=illust&illust_id={1}",base_url,illust_id)),
+                    String.Format("tt={0}&p=1&untagged=0&rest=show&book_id%5B%5D={1}&del=1",cookie_server.csrf_token,bookmark_id));
+                Console.WriteLine(ret);
+                return false;
+            }
+        }
+        //获取搜索结果
+        //!:key_word需要以URL编码
+        private async Task<List<int>> RequestSearchResult(string key_word, bool text_mode, int start_page, int end_page)
+        {
+            var ret = new List<int>();
+            try
+            {
+                var queue = new TaskQueue<List<int>>(25);
+                for (int i = start_page; i < end_page; ++i)//页数从1开始，在RequestSearchPage里面加1了
+                    await queue.Add(RequestSearchPage(key_word, i, text_mode));
+                await queue.Done();
+                foreach (var task in queue.done_task_list)
+                    ret.AddRange(task.Result);
+                Console.WriteLine("Search {0}:{1}_{2}", key_word.Substring(0, 20), ret.Count.ToString(), ret.Count > 0 ? ret[0].ToString() : "None");
+                return ret;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+            return ret;
+        }
+        public async Task<List<int>> RequestSearchPage(string word, int page, bool text_mode)
         {
             //s_mode:s_tc 在描述和标题里搜索 s_tag 在tag里搜索(部分一致) s_tag_full tag搜索(完全一致)
             //blt:最低收藏数 blg:最大收藏数
@@ -187,7 +310,7 @@ namespace PixivAss
                                     base_url,word, page+1,text_mode ? "s_tc" : "s_tag");
             string referer = String.Format("{0}tags/{1}/artworks?s_mode=s_tag_full", base_url,word);
             JObject json =await RequestJsonAsync(url, referer);
-            List<string> ret = new List<string>();
+            var ret = new List<int>();
             //排除包含非法关键字的图片
             foreach (var ill in json.GetValue("body").Value<JObject>().Value<JObject>("illustManga").Value<JArray>("data")) //这里返回的illust信息不全
             {
@@ -198,11 +321,46 @@ namespace PixivAss
                         if (this.banned_keyword.Contains(tag.ToString()))
                             valid = false;
                 if(valid)
-                    ret.Add(ill.ToObject<JObject>().Value<string>("illustId"));
+                    ret.Add(ill.ToObject<JObject>().Value<int>("illustId"));
             }
             return ret;
         }
-        public async Task<Illust> RequestIllustAsync(string illustId)
+        //返回<结果,总数>
+        public async Task<List<int>> RequestRankPage(string mode, int page)
+        {
+            //mode:daily weekly monthly male(受男性欢迎) female original(原创) rookie(新人)
+            //仅有daily/weekly/male/female可以带_r18后缀
+            //date:指定日期，形如20200921，没有获知过往排行的必要所以不使用
+            string url = String.Format("{0}ranking.php?mode={1}&p={2}&format=json",
+                                    base_url, mode, page + 1);
+            JObject json = await RequestJsonAsync(url, base_url);
+            var ret = new List<int>();
+            //排除包含非法关键字的图片
+            if(json.GetValue("contents")!=null)
+                foreach (var illust_object in json.Value<JArray>("contents"))
+                {
+                    var illust_id = illust_object.Value<int>("illust_id");
+                    var valid = true;
+                    var tags = illust_object.Value<JArray>("tags");
+                    if (tags != null)
+                        foreach (var tag in tags)
+                            if (this.banned_keyword.Contains(tag.ToString()))
+                                valid = false;
+                    //额外有一个
+                    var types = illust_object.Value<JObject>("illust_content_type");
+                    if (types.Value<bool>("homosexual") || types.Value<bool>("bl"))
+                        valid = false;
+                    if (valid)
+                        ret.Add(illust_id);
+                }
+            /*目前没有必要获取总数
+            int total = 0;
+            if (json.GetValue("rank_total")!=null)
+                total = json.Value<Int32>("rank_total");
+            */
+            return ret;
+        }
+        public async Task<Illust> RequestIllustAsync(int illustId)
         {
             string url = String.Format("{0}ajax/illust/{1}", base_url, illustId);
             string referer = String.Format("{0}member_illust.php?mode=medium&illust_id={1}", base_url, user_id);            
@@ -213,6 +371,5 @@ namespace PixivAss
                 return new Illust(illustId, false);
             return new Illust(json.Value<JObject>("body"));
         }
-
     }
 }
